Command heartbeats

+++ castleshooter/game.py
@@ -199,8 +199,8 @@ class Game:
                                         mouse_x, mouse_y = pygame.mouse.get_pos()
                                         unit_vector_from_player_to_mouse = get_unit_vector_from_player_to_mouse(client_player.x - x_offset, client_player.y - y_offset, mouse_x, mouse_y)
                                         arrow_distance = 400
-                                        arrow_dest_x = client_player.x + unit_vector_from_player_to_mouse[0] * arrow_distance
-                                        arrow_dest_y = client_player.y + unit_vector_from_player_to_mouse[1] * arrow_distance
+                                        arrow_dest_x = client_player.x + int(unit_vector_from_player_to_mouse[0] * arrow_distance)
+                                        arrow_dest_y = client_player.y + int(unit_vector_from_player_to_mouse[1] * arrow_distance)
                                         send_spawn_projectile_command(self.s, generate_projectile_id(), client_player.x, client_player.y, arrow_dest_x, arrow_dest_y, 
                                                                     [client.id, *[get_client_id_from_player_number(player_number, client_id=client.id) 
                                                                     for player_number in self.player_numbers_to_putative_teams.keys()
